[refactor] dashboard: 옷 통계 조회 API 전체 옷 수량 타입 int -> long 수정

+++ src/main/java/org/example/ootoutfitoftoday/domain/dashboard/dto/response/AdminClothesStatisticsResponse.java
@@ -12,7 +12,7 @@ import java.util.List;
 @AllArgsConstructor
 public class AdminClothesStatisticsResponse {
 
-    private final int totalClothes;
+    private final long totalClothes;
     private final List<CategoryStat> categoryStats;
     private final List<CountClothesColor> colorStats;
     private final List<CountClothesSize> sizeStats;
